refactor: split legacy-html constants
