+++ src/constants.mjs
@@ -7,17 +7,6 @@ export const DOC_NODE_VERSION = process.version;
 export const DOC_NODE_CHANGELOG_URL =
   'https://raw.githubusercontent.com/nodejs/node/HEAD/CHANGELOG.md';
 
-// This is the Node.js Base URL for viewing a file within GitHub UI
-export const DOC_NODE_BLOB_BASE_URL =
-  'https://github.com/nodejs/node/blob/HEAD/';
-
-// This is the Node.js API docs base URL for editing a file on GitHub UI
-export const DOC_API_BLOB_EDIT_BASE_URL =
-  'https://github.com/nodejs/node/edit/main/doc/api/';
-
-// Base URL for a specific Node.js version within the Node.js API docs
-export const DOC_API_BASE_URL_VERSION = 'https://nodejs.org/docs/latest-v';
-
 // This is the perma-link within the API docs that reference the Stability Index
 export const DOC_API_STABILITY_SECTION_REF_URL =
   'documentation.html#stability-index';
